Cleaned up Router API/Implementation
Rather than getting into the business of executing a handler (partly for
its side-effects), keeping the router free from that.  It just returns
the params it's matched in the route + the handler.

The HTTP server can handle the job of actually executing that handler,
and constructing the full Request object for the handler.