Added queryString support + generative tests
It does nothing but split at the first ?, and storing that as the QS.
Parsing of that can be done by middleware - since *how* querystrings get
parsed depends very much on your own environment.  Some teams use php
style qs (with the 'a[]=1&a[]=2' for arrays, iirc), whereas others just
repeat those (e.g. 'a=1&a=2')